# notes/stylenest-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/orders/{id}/cancel": {
            "put": {
                "tags": [
                    "order-controller"
                ],
                "operationId": "cancelOrder",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/payments/easebuzz/initiate": {
            "post": {
                "tags": [
                    "payment-controller"
                ],
                "operationId": "initiate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EasebuzzInitiateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseEasebuzzInitiateResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/payments/easebuzz/guest/initiate": {
            "post": {
                "tags": [
                    "payment-controller"
                ],
                "operationId": "initiateForGuest",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GuestPaymentInitiateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseEasebuzzInitiateResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/payments/easebuzz/callback": {
            "post": {
                "tags": [
                    "payment-controller"
                ],
                "operationId": "callback",
                "parameters": [
                    {
                        "name": "responseFields",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "order-controller"
                ],
                "operationId": "getMyOrders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListOrderSummaryResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "order-controller"
                ],
                "operationId": "placeOrder",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseOrderResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/guest/orders": {
            "post": {
                "tags": [
                    "guest-order-controller"
                ],
                "operationId": "placeOrder_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GuestOrderRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseOrderResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "order-controller"
                ],
                "operationId": "getOrderById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseOrderResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/{id}/invoice": {
            "get": {
                "tags": [
                    "order-controller"
                ],
                "operationId": "getInvoice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseInvoiceResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/{id}/invoice/pdf": {
            "get": {
                "tags": [
                    "order-controller"
                ],
                "operationId": "getInvoicePdf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "byte"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/guest/orders/invoice": {
            "get": {
                "tags": [
                    "guest-order-controller"
                ],
                "operationId": "getInvoice_1",
                "parameters": [
                    {
                        "name": "orderNumber",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "byte"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/orders/{id}": {
            "get": {
                "tags": [
                    "admin-order-controller"
                ],
                "operationId": "getOrderById_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseOrderResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/orders/{id}/invoice": {
            "get": {
                "tags": [
                    "admin-order-controller"
                ],
                "operationId": "getInvoice_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseInvoiceResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/orders/{id}/invoice/pdf": {
            "get": {
                "tags": [
                    "admin-order-controller"
                ],
                "operationId": "getInvoicePdf_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "byte"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResponseVoid": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {}
                }
            },
...
            "ApiResponseOrderResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/OrderResponse"
                    }
                }
            },
            "OrderItemResponse": {
                "type": "object",
                "properties": {
                    "productId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "productVariantId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "productName": {
                        "type": "string"
                    },
                    "color": {
                        "type": "string"
                    },
                    "size": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "price": {
                        "type": "number"
                    },
                    "subtotal": {
                        "type": "number"
                    }
                }
            },
            "OrderResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "orderNumber": {
                        "type": "string"
                    },
                    "totalAmount": {
                        "type": "number"
                    },
                    "paymentMethod": {
                        "type": "string",
                        "enum": [
                            "COD",
                            "CARD",
                            "UPI",
                            "NETBANKING"
                        ]
                    },
                    "paymentStatus": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "PAID",
                            "FAILED",
                            "REFUNDED",
                            "SUCCESS"
                        ]
                    },
                    "orderStatus": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "CONFIRMED",
                            "PROCESSING",
                            "COMPLETED",
                            "CANCELLED",
                            "PACKED",
                            "SHIPPED",
                            "DELIVERED"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItemResponse"
                        }
                    },
                    "isGuest": {
                        "type": "boolean"
                    },
                    "shippingAddress": {
                        "$ref": "#/components/schemas/ShippingAddressSnapshotResponse"
                    },
                    "shipmentStatus": {
                        "type": "string",
                        "enum": [
                            "PROCESSING",
                            "PACKED",
                            "SHIPPED",
                            "IN_TRANSIT",
                            "OUT_FOR_DELIVERY",
                            "DELIVERED",
                            "CANCELLED",
                            "RETURNED"
                        ]
                    },
                    "trackingNumber": {
                        "type": "string"
                    },
                    "courierName": {
                        "type": "string"
                    },
                    "estimatedDeliveryDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "invoiceAvailable": {
                        "type": "boolean"
                    }
                }
            },
            "ShippingAddressSnapshotResponse": {
                "type": "object",
                "properties": {
                    "fullName": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "phoneCountryCode": {
                        "type": "string"
                    },
                    "addressLine1": {
                        "type": "string"
                    },
                    "addressLine2": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "postalCode": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "countryCode": {
                        "type": "string"
                    }
                }
            },
...
            "EasebuzzInitiateRequest": {
                "type": "object",
                "properties": {
                    "paymentMethod": {
                        "type": "string",
                        "enum": [
                            "COD",
                            "CARD",
                            "UPI",
                            "NETBANKING"
                        ]
                    },
                    "upiVa": {
                        "type": "string"
                    },
                    "bankCode": {
                        "type": "string"
                    }
                },
                "required": [
                    "paymentMethod"
                ]
            },
            "ApiResponseEasebuzzInitiateResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/EasebuzzInitiateResponse"
                    }
                }
            },
            "EasebuzzInitiateResponse": {
                "type": "object",
                "properties": {
                    "orderId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "orderNumber": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "paymentMethod": {
                        "type": "string",
                        "enum": [
                            "COD",
                            "CARD",
                            "UPI",
                            "NETBANKING"
                        ]
                    },
                    "redirectUrl": {
                        "type": "string"
                    },
                    "bankRedirectHtml": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
            "GuestOrderItemRequest": {
                "type": "object",
                "properties": {
                    "productVariantId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "quantity": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    }
                },
                "required": [
                    "productVariantId",
                    "quantity"
                ]
            },
            "GuestPaymentInitiateRequest": {
                "type": "object",
                "properties": {
                    "guestEmail": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1
                    },
                    "shippingAddress": {
                        "$ref": "#/components/schemas/GuestShippingAddressRequest"
                    },
                    "paymentMethod": {
                        "type": "string",
                        "enum": [
                            "COD",
                            "CARD",
                            "UPI",
                            "NETBANKING"
                        ]
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GuestOrderItemRequest"
                        },
                        "minItems": 1
                    },
                    "upiVa": {
                        "type": "string"
                    },
                    "bankCode": {
                        "type": "string"
                    }
                },
                "required": [
                    "guestEmail",
                    "items",
                    "paymentMethod",
                    "shippingAddress"
                ]
            },
            "GuestShippingAddressRequest": {
                "type": "object",
                "properties": {
                    "fullName": {
                        "type": "string",
                        "minLength": 1
                    },
                    "phone": {
                        "type": "string",
                        "minLength": 1
                    },
                    "phoneCountryCode": {
                        "type": "string"
                    },
                    "addressLine1": {
                        "type": "string",
                        "minLength": 1
                    },
                    "addressLine2": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string",
                        "minLength": 1
                    },
                    "state": {
                        "type": "string",
                        "minLength": 1
                    },
                    "postalCode": {
                        "type": "string",
                        "minLength": 1
                    },
                    "country": {
                        "type": "string",
                        "minLength": 1
                    },
                    "countryCode": {
                        "type": "string"
                    }
                },
                "required": [
                    "addressLine1",
                    "city",
                    "country",
                    "fullName",
                    "phone",
                    "postalCode",
                    "state"
                ]
            },
            "OrderRequest": {
                "type": "object",
                "properties": {
                    "paymentMethod": {
                        "type": "string",
                        "enum": [
                            "COD",
                            "CARD",
                            "UPI",
                            "NETBANKING"
                        ]
                    }
                },
                "required": [
                    "paymentMethod"
                ]
            },
            "GuestOrderRequest": {
                "type": "object",
                "properties": {
                    "guestEmail": {
                        "type": "string",
                        "format": "email",
                        "minLength": 1
                    },
                    "shippingAddress": {
                        "$ref": "#/components/schemas/GuestShippingAddressRequest"
                    },
                    "paymentMethod": {
                        "type": "string",
                        "enum": [
                            "COD",
                            "CARD",
                            "UPI",
                            "NETBANKING"
                        ]
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/GuestOrderItemRequest"
                        },
                        "minItems": 1
                    }
                },
                "required": [
                    "guestEmail",
                    "items",
                    "paymentMethod",
                    "shippingAddress"
                ]
            },
...
            "ApiResponseListOrderSummaryResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderSummaryResponse"
                        }
                    }
                }
            },
            "OrderSummaryResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "orderNumber": {
                        "type": "string"
                    },
                    "totalAmount": {
                        "type": "number"
                    },
                    "orderStatus": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "CONFIRMED",
                            "PROCESSING",
                            "COMPLETED",
                            "CANCELLED",
                            "PACKED",
                            "SHIPPED",
                            "DELIVERED"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ApiResponseInvoiceResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/InvoiceResponse"
                    }
                }
            },
...
        }
    }
}